moved limelight enums to their own files
